@@ -14,12 +14,22 @@ declare(strict_types=1);
 
 namespace FOD\DBALClickHouse;
 
+use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\Exception\InvalidColumnDeclaration;
+use Doctrine\DBAL\Exception\InvalidColumnType;
+use Doctrine\DBAL\LockMode;
 use Doctrine\DBAL\ParameterType;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
+use Doctrine\DBAL\Platforms\DateIntervalUnit;
+use Doctrine\DBAL\Platforms\Exception\NotSupported;
+use Doctrine\DBAL\Platforms\Keywords\KeywordList;
 use Doctrine\DBAL\Platforms\TrimMode;
+use Doctrine\DBAL\Schema\AbstractSchemaManager;
 use Doctrine\DBAL\Schema\ForeignKeyConstraint;
 use Doctrine\DBAL\Schema\Index;
 use Doctrine\DBAL\Schema\TableDiff;
+use Doctrine\DBAL\Schema\UniqueConstraint;
+use Doctrine\DBAL\TransactionIsolationLevel;
 use Doctrine\DBAL\Types\BigIntType;
 use Doctrine\DBAL\Types\BlobType;
 use Doctrine\DBAL\Types\DateTimeType;
@@ -34,8 +44,9 @@ use Doctrine\DBAL\Types\Type;
 use FOD\DBALClickHouse\Types\BitNumericalClickHouseType;
 use FOD\DBALClickHouse\Types\DatableClickHouseType;
 use FOD\DBALClickHouse\Types\NumericalClickHouseType;
-use FOD\DBALClickHouse\Types\StringClickHouseType;
+use FOD\DBALClickHouse\Types\StringableClickHouseType;
 use FOD\DBALClickHouse\Types\UnsignedNumericalClickHouseType;
+
 use function addslashes;
 use function array_filter;
 use function array_key_exists;
@@ -44,149 +55,140 @@ use function array_merge;
 use function array_unique;
 use function array_values;
 use function count;
-use function func_get_args;
 use function get_class;
 use function implode;
 use function in_array;
+use function mb_stripos;
 use function sprintf;
-use function stripos;
 use function trim;
 
-/**
- * Provides the behavior, features and SQL dialect of the ClickHouse database platform.
- */
 class ClickHousePlatform extends AbstractPlatform
 {
-    protected const TIME_MINUTE = 60;
-    protected const TIME_HOUR   = self::TIME_MINUTE * 60;
-    protected const TIME_DAY    = self::TIME_HOUR * 24;
-    protected const TIME_WEEK   = self::TIME_DAY * 7;
-
     /**
      * {@inheritDoc}
      */
-    public function getBooleanTypeDeclarationSQL(array $columnDef) : string
+    public function getBooleanTypeDeclarationSQL(array $column): string
     {
         return $this->prepareDeclarationSQL(
             UnsignedNumericalClickHouseType::UNSIGNED_CHAR .
             NumericalClickHouseType::TYPE_INT . BitNumericalClickHouseType::EIGHT_BIT,
-            $columnDef
+            $column
         );
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getIntegerTypeDeclarationSQL(array $columnDef) : string
+    public function getIntegerTypeDeclarationSQL(array $column): string
     {
         return $this->prepareDeclarationSQL(
-            $this->_getCommonIntegerTypeDeclarationSQL($columnDef) .
+            $this->_getCommonIntegerTypeDeclarationSQL($column) .
             NumericalClickHouseType::TYPE_INT . BitNumericalClickHouseType::THIRTY_TWO_BIT,
-            $columnDef
+            $column
         );
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getBigIntTypeDeclarationSQL(array $columnDef) : string
+    public function getBigIntTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(StringClickHouseType::TYPE_STRING, $columnDef);
+        return $this->prepareDeclarationSQL(StringableClickHouseType::TYPE_STRING, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getSmallIntTypeDeclarationSQL(array $columnDef) : string
+    public function getSmallIntTypeDeclarationSQL(array $column): string
     {
         return $this->prepareDeclarationSQL(
-            $this->_getCommonIntegerTypeDeclarationSQL($columnDef) .
+            $this->_getCommonIntegerTypeDeclarationSQL($column) .
             NumericalClickHouseType::TYPE_INT . BitNumericalClickHouseType::SIXTEEN_BIT,
-            $columnDef
+            $column
         );
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function _getCommonIntegerTypeDeclarationSQL(array $columnDef) : string
+    protected function _getCommonIntegerTypeDeclarationSQL(array $column): string
     {
-        if (! empty($columnDef['autoincrement'])) {
-            throw ClickHouseException::notSupported('Clickhouse do not support AUTO_INCREMENT fields');
+        if (!empty($column['autoincrement'])) {
+            throw NotSupported::new('Clickhouse does not support AUTO_INCREMENT fields');
         }
 
-        return empty($columnDef['unsigned']) ? '' : UnsignedNumericalClickHouseType::UNSIGNED_CHAR;
+        return empty($column['unsigned']) ? '' : UnsignedNumericalClickHouseType::UNSIGNED_CHAR;
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function initializeDoctrineTypeMappings() : void
+    protected function initializeDoctrineTypeMappings(): void
     {
         $this->doctrineTypeMapping = [
-            'int8' => 'smallint',
-            'int16' => 'integer',
-            'int32' => 'integer',
-            'int64' => 'bigint',
-            'uint8' => 'smallint',
-            'uint16' => 'integer',
-            'uint32' => 'integer',
-            'uint64' => 'bigint',
+            'int8'    => 'smallint',
+            'int16'   => 'integer',
+            'int32'   => 'integer',
+            'int64'   => 'bigint',
+            'uint8'   => 'smallint',
+            'uint16'  => 'integer',
+            'uint32'  => 'integer',
+            'uint64'  => 'bigint',
             'float32' => 'decimal',
             'float64' => 'float',
 
-            'string' => 'string',
+            'string'      => 'string',
             'fixedstring' => 'string',
-            'date' => 'date',
-            'datetime' => 'datetime',
+            'date'        => 'date',
+            'datetime'    => 'datetime',
 
-            'array(int8)' => 'array',
-            'array(int16)' => 'array',
-            'array(int32)' => 'array',
-            'array(int64)' => 'array',
-            'array(uint8)' => 'array',
-            'array(uint16)' => 'array',
-            'array(uint32)' => 'array',
-            'array(uint64)' => 'array',
+            'array(int8)'    => 'array',
+            'array(int16)'   => 'array',
+            'array(int32)'   => 'array',
+            'array(int64)'   => 'array',
+            'array(uint8)'   => 'array',
+            'array(uint16)'  => 'array',
+            'array(uint32)'  => 'array',
+            'array(uint64)'  => 'array',
             'array(float32)' => 'array',
             'array(float64)' => 'array',
 
-            'array(string)' => 'array',
-            'array(date)' => 'array',
+            'array(string)'   => 'array',
+            'array(date)'     => 'array',
             'array(datetime)' => 'array',
 
-            'enum8' => 'string',
+            'enum8'  => 'string',
             'enum16' => 'string',
 
-            'nullable(int8)' => 'smallint',
-            'nullable(int16)' => 'integer',
-            'nullable(int32)' => 'integer',
-            'nullable(int64)' => 'bigint',
-            'nullable(uint8)' => 'smallint',
-            'nullable(uint16)' => 'integer',
-            'nullable(uint32)' => 'integer',
-            'nullable(uint64)' => 'bigint',
+            'nullable(int8)'    => 'smallint',
+            'nullable(int16)'   => 'integer',
+            'nullable(int32)'   => 'integer',
+            'nullable(int64)'   => 'bigint',
+            'nullable(uint8)'   => 'smallint',
+            'nullable(uint16)'  => 'integer',
+            'nullable(uint32)'  => 'integer',
+            'nullable(uint64)'  => 'bigint',
             'nullable(float32)' => 'decimal',
             'nullable(float64)' => 'float',
 
-            'nullable(string)' => 'string',
+            'nullable(string)'      => 'string',
             'nullable(fixedstring)' => 'string',
-            'nullable(date)' => 'date',
-            'nullable(datetime)' => 'datetime',
+            'nullable(date)'        => 'date',
+            'nullable(datetime)'    => 'datetime',
 
-            'array(nullable(int8))' => 'array',
-            'array(nullable(int16))' => 'array',
-            'array(nullable(int32))' => 'array',
-            'array(nullable(int64))' => 'array',
-            'array(nullable(uint8))' => 'array',
-            'array(nullable(uint16))' => 'array',
-            'array(nullable(uint32))' => 'array',
-            'array(nullable(uint64))' => 'array',
+            'array(nullable(int8))'    => 'array',
+            'array(nullable(int16))'   => 'array',
+            'array(nullable(int32))'   => 'array',
+            'array(nullable(int64))'   => 'array',
+            'array(nullable(uint8))'   => 'array',
+            'array(nullable(uint16))'  => 'array',
+            'array(nullable(uint32))'  => 'array',
+            'array(nullable(uint64))'  => 'array',
             'array(nullable(float32))' => 'array',
             'array(nullable(float64))' => 'array',
 
-            'array(nullable(string))' => 'array',
-            'array(nullable(date))' => 'array',
+            'array(nullable(string))'   => 'array',
+            'array(nullable(date))'     => 'array',
             'array(nullable(datetime))' => 'array',
         ];
     }
@@ -194,92 +196,76 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    protected function getVarcharTypeDeclarationSQLSnippet($length, $fixed) : string
+    protected function getVarcharTypeDeclarationSQLSnippet(?int $length): string
     {
-        return $fixed
-            ? (StringClickHouseType::TYPE_FIXED_STRING . '(' . $length . ')')
-            : StringClickHouseType::TYPE_STRING;
+        return StringableClickHouseType::TYPE_STRING;
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function getVarcharTypeDeclarationSQL(array $field) : string
+    public function getStringTypeDeclarationSQL(array $column): string
     {
-        if (! isset($field['length'])) {
-            $field['length'] = $this->getVarcharDefaultLength();
+        $length = $column['length'] ?? null;
+
+        if (empty($column['fixed'])) {
+            try {
+                return $this->prepareDeclarationSQL($this->getVarcharTypeDeclarationSQLSnippet($length), $column);
+            } catch (InvalidColumnType $e) {
+                throw InvalidColumnDeclaration::fromInvalidColumnType($column['name'], $e);
+            }
         }
 
-        $fixed = $field['fixed'] ?? false;
+        return $this->prepareDeclarationSQL($this->getCharTypeDeclarationSQLSnippet($length), $column);
+    }
 
-        $maxLength = $fixed
-            ? $this->getCharMaxLength()
-            : $this->getVarcharMaxLength();
+    /**
+     * @param int|null $length The length of the column in characters
+     *                         or NULL if the length should be omitted.
+     */
+    protected function getCharTypeDeclarationSQLSnippet(?int $length): string
+    {
+        $sql = StringableClickHouseType::TYPE_FIXED_STRING;
 
-        if ($field['length'] > $maxLength) {
-            return $this->getClobTypeDeclarationSQL($field);
+        if ($length !== null) {
+            $sql .= sprintf('(%d)', $length);
         }
 
-        return $this->prepareDeclarationSQL(
-            $this->getVarcharTypeDeclarationSQLSnippet($field['length'], $fixed),
-            $field
-        );
+        return $sql;
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function getBinaryTypeDeclarationSQLSnippet($length, $fixed) : string
+    protected function getBinaryTypeDeclarationSQLSnippet(?int $length): string
     {
-        return StringClickHouseType::TYPE_STRING;
+        if (null === $length) {
+            return StringableClickHouseType::TYPE_STRING;
+        }
+
+        return StringableClickHouseType::TYPE_FIXED_STRING . '(' . $length . ')';
+    }
+
+    protected function getVarbinaryTypeDeclarationSQLSnippet(?int $length): string
+    {
+        if (null === $length) {
+            return StringableClickHouseType::TYPE_STRING;
+        }
+
+        return StringableClickHouseType::TYPE_FIXED_STRING . '(' . $length . ')';
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getClobTypeDeclarationSQL(array $field) : string
+    public function getClobTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(StringClickHouseType::TYPE_STRING, $field);
+        return $this->prepareDeclarationSQL(StringableClickHouseType::TYPE_STRING, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getBlobTypeDeclarationSQL(array $field) : string
+    public function getBlobTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(StringClickHouseType::TYPE_STRING, $field);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getName() : string
-    {
-        return 'clickhouse';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getIdentifierQuoteCharacter() : string
-    {
-        return '`';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getVarcharDefaultLength() : int
-    {
-        return 512;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getCountExpression($column) : string
-    {
-        return 'COUNT()';
+        return $this->prepareDeclarationSQL(StringableClickHouseType::TYPE_STRING, $column);
     }
 
     // scalar functions
@@ -287,57 +273,30 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getMd5Expression($column) : string
+    public function getLengthExpression(string $string): string
     {
-        return 'MD5(CAST(' . $column . ' AS String))';
+        return 'lengthUTF8(CAST(' . $string . ' AS String))';
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getLengthExpression($column) : string
+    public function getModExpression(string $dividend, string $divisor): string
     {
-        return 'lengthUTF8(CAST(' . $column . ' AS String))';
+        return 'modulo(' . $dividend . ', ' . $divisor . ')';
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getSqrtExpression($column) : string
+    public function getTrimExpression(string $str, TrimMode $mode = TrimMode::UNSPECIFIED, ?string $char = null): string
     {
-        return 'sqrt(' . $column . ')';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getRoundExpression($column, $decimals = 0) : string
-    {
-        return 'round(' . $column . ', ' . $decimals . ')';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getModExpression($expression1, $expression2) : string
-    {
-        return 'modulo(' . $expression1 . ', ' . $expression2 . ')';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getTrimExpression($str, $pos = TrimMode::UNSPECIFIED, $char = false) : string
-    {
-        if (! $char) {
-            switch ($pos) {
-                case TrimMode::LEADING:
-                    return $this->getLtrimExpression($str);
-                case TrimMode::TRAILING:
-                    return $this->getRtrimExpression($str);
-                default:
-                    return sprintf("replaceRegexpAll(%s, '(^\\\s+|\\\s+$)', '')", $str);
-            }
+        if (null === $char) {
+            return match ($mode) {
+                TrimMode::LEADING => sprintf("replaceRegexpAll(%s, '(^\\\s+)', '')", $str),
+                TrimMode::TRAILING => sprintf("replaceRegexpAll(%s, '(\\\s+$)', '')", $str),
+                default => sprintf("replaceRegexpAll(%s, '(^\\\s+|\\\s+$)', '')", $str),
+            };
         }
 
         return sprintf("replaceRegexpAll(%s, '(^%s+|%s+$)', '')", $str, $char, $char);
@@ -346,123 +305,35 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getRtrimExpression($str) : string
+    public function getLocateExpression(string $string, string $substring, ?string $start = null): string
     {
-        return sprintf("replaceRegexpAll(%s, '(\\\s+$)', '')", $str);
+        return 'positionUTF8(' . $string . ', ' . $substring . ')';
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getLtrimExpression($str) : string
-    {
-        return sprintf("replaceRegexpAll(%s, '(^\\\s+)', '')", $str);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getUpperExpression($str) : string
-    {
-        return 'upperUTF8(' . $str . ')';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getLowerExpression($str) : string
-    {
-        return 'lowerUTF8(' . $str . ')';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getLocateExpression($str, $substr, $startPos = false) : string
-    {
-        return 'positionUTF8(' . $str . ', ' . $substr . ')';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getNowExpression() : string
-    {
-        return 'now()';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getSubstringExpression($value, $from, $length = null) : string
+    public function getSubstringExpression(string $string, string $start, ?string $length = null): string
     {
         if ($length === null) {
             throw new \InvalidArgumentException("'length' argument must be a constant");
         }
 
-        return 'substringUTF8(' . $value . ', ' . $from . ', ' . $length . ')';
+        return 'substringUTF8(' . $string . ', ' . $start . ', ' . $length . ')';
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getConcatExpression() : string
+    public function getConcatExpression(string ...$string): string
     {
-        return 'concat(' . implode(', ', func_get_args()) . ')';
+        return 'concat(' . implode(', ', $string) . ')';
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getIsNullExpression($expression) : string
-    {
-        return 'isNull(' . $expression . ')';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getIsNotNullExpression($expression) : string
-    {
-        return 'isNotNull(' . $expression . ')';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getAcosExpression($value) : string
-    {
-        return 'acos(' . $value . ')';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getSinExpression($value) : string
-    {
-        return 'sin(' . $value . ')';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getPiExpression() : string
-    {
-        return 'pi()';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getCosExpression($value) : string
-    {
-        return 'cos(' . $value . ')';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getDateDiffExpression($date1, $date2) : string
+    public function getDateDiffExpression(string $date1, string $date2): string
     {
         return 'CAST(' . $date1 . ' AS Date) - CAST(' . $date2 . ' AS Date)';
     }
@@ -470,87 +341,87 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateAddSecondsExpression($date, $seconds) : string
+    public function getDateAddSecondsExpression(string $date, string $seconds): string
     {
-        return $date . ' + ' . $seconds;
+        return $this->getDateArithmeticIntervalExpression($date, '+', $seconds, DateIntervalUnit::SECOND);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateSubSecondsExpression($date, $seconds) : string
+    public function getDateSubSecondsExpression(string $date, string $seconds): string
     {
-        return $date . ' - ' . $seconds;
+        return $this->getDateArithmeticIntervalExpression($date, '-', $seconds, DateIntervalUnit::SECOND);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateAddMinutesExpression($date, $minutes) : string
+    public function getDateAddMinutesExpression(string $date, string $minutes): string
     {
-        return $date . ' + ' . $minutes * self::TIME_MINUTE;
+        return $this->getDateArithmeticIntervalExpression($date, '+', $minutes, DateIntervalUnit::MINUTE);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateSubMinutesExpression($date, $minutes) : string
+    public function getDateSubMinutesExpression(string $date, string $minutes): string
     {
-        return $date . ' - ' . $minutes * self::TIME_MINUTE;
+        return $this->getDateArithmeticIntervalExpression($date, '-', $minutes, DateIntervalUnit::MINUTE);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateAddHourExpression($date, $hours) : string
+    public function getDateAddHourExpression(string $date, string $hours): string
     {
-        return $date . ' + ' . $hours * self::TIME_HOUR;
+        return $this->getDateArithmeticIntervalExpression($date, '+', $hours, DateIntervalUnit::HOUR);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateSubHourExpression($date, $hours) : string
+    public function getDateSubHourExpression(string $date, string $hours): string
     {
-        return $date . ' - ' . $hours * self::TIME_HOUR;
+        return $this->getDateArithmeticIntervalExpression($date, '-', $hours, DateIntervalUnit::HOUR);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateAddDaysExpression($date, $days) : string
+    public function getDateAddDaysExpression(string $date, string $days): string
     {
-        return $date . ' + ' . $days * self::TIME_DAY;
+        return $this->getDateArithmeticIntervalExpression($date, '+', $days, DateIntervalUnit::DAY);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateSubDaysExpression($date, $days) : string
+    public function getDateSubDaysExpression(string $date, string $days): string
     {
-        return $date . ' - ' . $days * self::TIME_DAY;
+        return $this->getDateArithmeticIntervalExpression($date, '-', $days, DateIntervalUnit::DAY);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateAddWeeksExpression($date, $weeks) : string
+    public function getDateAddWeeksExpression(string $date, string $weeks): string
     {
-        return $date . ' + ' . $weeks * self::TIME_WEEK;
+        return $this->getDateArithmeticIntervalExpression($date, '+', $weeks, DateIntervalUnit::WEEK);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateSubWeeksExpression($date, $weeks) : string
+    public function getDateSubWeeksExpression(string $date, string $weeks): string
     {
-        return $date . ' - ' . $weeks * self::TIME_WEEK;
+        return $this->getDateArithmeticIntervalExpression($date, '-', $weeks, DateIntervalUnit::WEEK);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getBitAndComparisonExpression($value1, $value2) : string
+    public function getBitAndComparisonExpression(string $value1, string $value2): string
     {
         return 'bitAnd(' . $value1 . ', ' . $value2 . ')';
     }
@@ -558,7 +429,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getBitOrComparisonExpression($value1, $value2) : string
+    public function getBitOrComparisonExpression(string $value1, string $value2): string
     {
         return 'bitOr(' . $value1 . ', ' . $value2 . ')';
     }
@@ -566,81 +437,59 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getForUpdateSQL() : string
+    public function appendLockHint(string $fromClause, LockMode $lockMode): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function appendLockHint($fromClause, $lockMode) : string
+    public function getDropIndexSQL(string $name, string $table): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getReadLockSQL() : string
+    public function getDropConstraintSQL(string $name, string $table): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getWriteLockSQL() : string
+    public function getDropForeignKeySQL(string $foreignKey, string $table): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDropIndexSQL($index, $table = null) : string
+    public function getCommentOnColumnSQL(string $tableName, string $columnName, string $comment): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDropConstraintSQL($constraint, $table) : string
+    protected function _getCreateTableSQL(string $name, array $columns, array $options = []): array
     {
-        throw ClickHouseException::notSupported(__METHOD__);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getDropForeignKeySQL($foreignKey, $table) : string
-    {
-        throw ClickHouseException::notSupported(__METHOD__);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getCommentOnColumnSQL($tableName, $columnName, $comment) : string
-    {
-        throw ClickHouseException::notSupported(__METHOD__);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    protected function _getCreateTableSQL($tableName, array $columns, array $options = []) : array
-    {
-        $engine        = ! empty($options['engine']) ? $options['engine'] : 'ReplacingMergeTree';
+        $engine        = !empty($options['engine']) ? $options['engine'] : 'ReplacingMergeTree';
         $engineOptions = '';
 
-        if (isset($options['uniqueConstraints']) && ! empty($options['uniqueConstraints'])) {
-            throw ClickHouseException::notSupported('uniqueConstraints');
+        $columns = \array_column($columns, null, 'name');
+
+        if (isset($options['uniqueConstraints']) && !empty($options['uniqueConstraints'])) {
+            throw NotSupported::new('uniqueConstraints');
         }
 
-        if (isset($options['indexes']) && ! empty($options['indexes'])) {
-            throw ClickHouseException::notSupported('uniqueConstraints');
+        if (isset($options['indexes']) && !empty($options['indexes'])) {
+            throw NotSupported::new('uniqueConstraints');
         }
 
         /**
@@ -658,40 +507,42 @@ class ClickHousePlatform extends AbstractPlatform
             ],
             true
         )) {
-            $indexGranularity   = ! empty($options['indexGranularity']) ? $options['indexGranularity'] : 8192;
+            $indexGranularity   = !empty($options['indexGranularity']) ? $options['indexGranularity'] : 8192;
             $samplingExpression = '';
 
             /**
              * eventDateColumn section
              */
             $dateColumnParams = [
-                'type' => Type::getType('date'),
+                'type'    => Type::getType('date'),
                 'default' => 'today()',
             ];
-            if (! empty($options['eventDateProviderColumn'])) {
+
+            if (!empty($options['eventDateProviderColumn'])) {
                 $options['eventDateProviderColumn'] = trim($options['eventDateProviderColumn']);
-                if (! isset($columns[$options['eventDateProviderColumn']])) {
+
+                if (!isset($columns[$options['eventDateProviderColumn']])) {
                     throw new \Exception(
-                        'Table `' . $tableName . '` has not column with name: `' . $options['eventDateProviderColumn']
+                        'Table `' . $name . '` has not column with name: `' . $options['eventDateProviderColumn']
                     );
                 }
 
-                if (! ($columns[$options['eventDateProviderColumn']]['type'] instanceof DateType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof DateTimeType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof TextType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof IntegerType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof SmallIntType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof BigIntType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof FloatType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof DecimalType) &&
+                if (!($columns[$options['eventDateProviderColumn']]['type'] instanceof DateType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof DateTimeType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof TextType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof IntegerType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof SmallIntType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof BigIntType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof FloatType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof DecimalType) &&
                     (
-                        ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof StringType) ||
+                        !($columns[$options['eventDateProviderColumn']]['type'] instanceof StringType) ||
                         $columns[$options['eventDateProviderColumn']]['fixed']
                     )
                 ) {
                     throw new \Exception(
                         'Column `' . $options['eventDateProviderColumn'] . '` with type `' .
-                        $columns[$options['eventDateProviderColumn']]['type']->getName() .
+                        $columns[$options['eventDateProviderColumn']]['type']->lookupName($columns[$options['eventDateProviderColumn']]['type']) .
                         '`, defined in `eventDateProviderColumn` option, has not valid DBAL Type'
                     );
                 }
@@ -703,14 +554,13 @@ class ClickHousePlatform extends AbstractPlatform
                         ('toDate(toDateTime(' . $options['eventDateProviderColumn'] . '))') :
                         ('toDate(' . $options['eventDateProviderColumn'] . ')');
             }
+
             if (empty($options['eventDateColumn'])) {
-                $dateColumns = array_filter($columns, function ($column) {
-                    return $column['type'] instanceof DateType;
-                });
+                $dateColumns = array_filter($columns, fn (array $column): bool => $column['type'] instanceof DateType);
 
                 if ($dateColumns) {
                     throw new \Exception(
-                        'Table `' . $tableName . '` has DateType columns: `' . implode(
+                        'Table `' . $name . '` has DateType columns: `' . implode(
                             '`, `',
                             array_keys($dateColumns)
                         ) .
@@ -721,9 +571,9 @@ class ClickHousePlatform extends AbstractPlatform
 
                 $eventDateColumnName = 'EventDate';
             } elseif (isset($columns[$options['eventDateColumn']])) {
-                if (! ($columns[$options['eventDateColumn']]['type'] instanceof DateType)) {
+                if (!($columns[$options['eventDateColumn']]['type'] instanceof DateType)) {
                     throw new \Exception(
-                        'In table `' . $tableName . '` you have set field `' .
+                        'In table `' . $name . '` you have set field `' .
                         $options['eventDateColumn'] .
                         '` (' . get_class($columns[$options['eventDateColumn']]['type']) . ')
                          as `eventDateColumn`, but it is not instance of DateType'
@@ -731,10 +581,12 @@ class ClickHousePlatform extends AbstractPlatform
                 }
 
                 $eventDateColumnName = $options['eventDateColumn'];
+
                 unset($columns[$options['eventDateColumn']]);
             } else {
                 $eventDateColumnName = $options['eventDateColumn'];
             }
+
             $dateColumnParams['name'] = $eventDateColumnName;
             // insert into very beginning
             $columns = [$eventDateColumnName => $dateColumnParams] + $columns;
@@ -747,7 +599,8 @@ class ClickHousePlatform extends AbstractPlatform
             }
 
             $primaryIndex = array_values($options['primary']);
-            if (! empty($options['samplingExpression'])) {
+
+            if (!empty($options['samplingExpression'])) {
                 $samplingExpression = ', ' . $options['samplingExpression'];
                 $primaryIndex[]     = $options['samplingExpression'];
             }
@@ -766,19 +619,19 @@ class ClickHousePlatform extends AbstractPlatform
             /**
              * any specific MergeTree* table parameters
              */
-            if ($engine === 'ReplacingMergeTree' && ! empty($options['versionColumn'])) {
-                if (! isset($columns[$options['versionColumn']])) {
+            if ($engine === 'ReplacingMergeTree' && !empty($options['versionColumn'])) {
+                if (!isset($columns[$options['versionColumn']])) {
                     throw new \Exception(
                         'If you specify `versionColumn` for ReplacingMergeTree table -- 
                         you must add this column manually (any of UInt*, Date or DateTime types)'
                     );
                 }
 
-                if (! $columns[$options['versionColumn']]['type'] instanceof IntegerType &&
-                    ! $columns[$options['versionColumn']]['type'] instanceof BigIntType &&
-                    ! $columns[$options['versionColumn']]['type'] instanceof SmallIntType &&
-                    ! $columns[$options['versionColumn']]['type'] instanceof DateType &&
-                    ! $columns[$options['versionColumn']]['type'] instanceof DateTimeType
+                if (!$columns[$options['versionColumn']]['type'] instanceof IntegerType &&
+                    !$columns[$options['versionColumn']]['type'] instanceof BigIntType &&
+                    !$columns[$options['versionColumn']]['type'] instanceof SmallIntType &&
+                    !$columns[$options['versionColumn']]['type'] instanceof DateType &&
+                    !$columns[$options['versionColumn']]['type'] instanceof DateTimeType
                 ) {
                     throw new \Exception(
                         'For ReplacingMergeTree tables `versionColumn` must be any of UInt* family, 
@@ -795,7 +648,7 @@ class ClickHousePlatform extends AbstractPlatform
 
         $sql[] = sprintf(
             'CREATE TABLE %s (%s) ENGINE = %s%s',
-            $tableName,
+            $name,
             $this->getColumnDeclarationListSQL($columns),
             $engine,
             $engineOptions
@@ -807,69 +660,57 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCreateForeignKeySQL(ForeignKeyConstraint $foreignKey, $table) : string
+    public function getCreateForeignKeySQL(ForeignKeyConstraint $foreignKey, string $table): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getAlterTableSQL(TableDiff $diff) : array
+    public function getAlterTableSQL(TableDiff $diff): array
     {
         $columnSql  = [];
         $queryParts = [];
-        if ($diff->newName !== false || ! empty($diff->renamedColumns)) {
-            throw ClickHouseException::notSupported('RENAME COLUMN');
+
+        foreach ($diff->getChangedColumns() as $column) {
+            if ($column->getNewColumn()->getName() !== $column->getOldColumn()->getName()) {
+                throw NotSupported::new('RENAME COLUMN');
+            }
         }
 
-        foreach ($diff->addedColumns as $column) {
-            if ($this->onSchemaAlterTableAddColumn($column, $diff, $columnSql)) {
-                continue;
-            }
-
+        foreach ($diff->getAddedColumns() as $column) {
             $columnArray  = $column->toArray();
             $queryParts[] = 'ADD COLUMN ' . $this->getColumnDeclarationSQL($column->getQuotedName($this), $columnArray);
         }
 
-        foreach ($diff->removedColumns as $column) {
-            if ($this->onSchemaAlterTableRemoveColumn($column, $diff, $columnSql)) {
-                continue;
-            }
-
+        foreach ($diff->getDroppedColumns() as $column) {
             $queryParts[] = 'DROP COLUMN ' . $column->getQuotedName($this);
         }
 
-        foreach ($diff->changedColumns as $columnDiff) {
-            if ($this->onSchemaAlterTableChangeColumn($columnDiff, $diff, $columnSql)) {
-                continue;
-            }
-
-            $column      = $columnDiff->column;
+        foreach ($diff->getChangedColumns() as $columnDiff) {
+            $column      = $columnDiff->getNewColumn();
             $columnArray = $column->toArray();
 
             // Don't propagate default value changes for unsupported column types.
-            if (($columnArray['type'] instanceof TextType || $columnArray['type'] instanceof BlobType) &&
-                $columnDiff->hasChanged('default') &&
-                count($columnDiff->changedProperties) === 1
+            if (($columnArray['type'] instanceof TextType || $columnArray['type'] instanceof BlobType)
+                && $columnDiff->hasDefaultChanged()
+                && $columnDiff->countChangedProperties() === 1
             ) {
                 continue;
             }
 
             $queryParts[] = 'MODIFY COLUMN ' . $this->getColumnDeclarationSQL(
-                $column->getQuotedName($this),
-                $columnArray
-            );
+                    $column->getQuotedName($this),
+                    $columnArray
+                );
         }
 
         $sql      = [];
         $tableSql = [];
 
-        if (! $this->onSchemaAlterTable($diff, $tableSql) && (count($queryParts) > 0)) {
-            $sql[] = 'ALTER TABLE ' . $diff->getName($this)->getQuotedName($this) . ' ' . implode(
-                ', ',
-                $queryParts
-            );
+        if ((count($queryParts) > 0)) {
+            $sql[] = 'ALTER TABLE ' . $diff->getOldTable()->getQuotedName($this) . ' ' . implode(', ', $queryParts);
         }
 
         return array_merge($sql, $tableSql, $columnSql);
@@ -878,36 +719,28 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    protected function getPreAlterTableIndexForeignKeySQL(TableDiff $diff) : array
+    protected function getPreAlterTableIndexForeignKeySQL(TableDiff $diff): array
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function getPostAlterTableIndexForeignKeySQL(TableDiff $diff) : array
+    protected function getPostAlterTableIndexForeignKeySQL(TableDiff $diff): array
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function getRenameIndexSQL($oldIndexName, Index $index, $tableName) : array
+    protected function getRenameIndexSQL(string $oldIndexName, Index $index, string $tableName): array
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    protected function _getAlterTableIndexForeignKeySQL(TableDiff $diff) : array
-    {
-        throw ClickHouseException::notSupported(__METHOD__);
-    }
-
-    protected function prepareDeclarationSQL(string $declarationSQL, array $columnDef) : string
+    protected function prepareDeclarationSQL(string $declarationSQL, array $columnDef): string
     {
         if (array_key_exists('notnull', $columnDef) && $columnDef['notnull'] === false) {
             return 'Nullable(' . $declarationSQL . ')';
@@ -919,14 +752,14 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getColumnDeclarationSQL($name, array $field) : string
+    public function getColumnDeclarationSQL(string $name, array $column): string
     {
-        if (isset($field['columnDefinition'])) {
-            $columnDef = $this->getCustomTypeDeclarationSQL($field);
+        if (isset($column['columnDefinition'])) {
+            $columnDef = $column['columnDefinition'];
         } else {
-            $default = $this->getDefaultValueDeclarationSQL($field);
+            $default = $this->getDefaultValueDeclarationSQL($column);
 
-            $columnDef = $field['type']->getSqlDeclaration($field, $this) . $default;
+            $columnDef = $column['type']->getSqlDeclaration($column, $this) . $default;
         }
 
         return $name . ' ' . $columnDef;
@@ -935,80 +768,71 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDecimalTypeDeclarationSQL(array $columnDef) : string
+    public function getDecimalTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(StringClickHouseType::TYPE_STRING, $columnDef);
+        return $this->prepareDeclarationSQL(StringableClickHouseType::TYPE_STRING, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getCheckDeclarationSQL(array $definition) : string
+    public function getCheckDeclarationSQL(array $definition): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getUniqueConstraintDeclarationSQL($name, Index $index) : string
+    public function getUniqueConstraintDeclarationSQL(UniqueConstraint $constraint): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getIndexDeclarationSQL($name, Index $index) : string
+    public function getIndexDeclarationSQL(Index $index): string
     {
-        // Index declaration in statements like CREATE TABLE is not supported.
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getForeignKeyDeclarationSQL(ForeignKeyConstraint $foreignKey) : string
+    public function getForeignKeyDeclarationSQL(ForeignKeyConstraint $foreignKey): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getAdvancedForeignKeyOptionsSQL(ForeignKeyConstraint $foreignKey) : string
+    public function getAdvancedForeignKeyOptionsSQL(ForeignKeyConstraint $foreignKey): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getForeignKeyReferentialActionSQL($action) : string
+    public function getForeignKeyReferentialActionSQL(string $action): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getForeignKeyBaseDeclarationSQL(ForeignKeyConstraint $foreignKey) : string
+    public function getForeignKeyBaseDeclarationSQL(ForeignKeyConstraint $foreignKey): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getUniqueFieldDeclarationSQL() : string
-    {
-        throw ClickHouseException::notSupported(__METHOD__);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getCurrentDateSQL() : string
+    public function getCurrentDateSQL(): string
     {
         return 'today()';
     }
@@ -1016,7 +840,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCurrentTimeSQL() : string
+    public function getCurrentTimeSQL(): string
     {
         return 'now()';
     }
@@ -1024,7 +848,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCurrentTimestampSQL() : string
+    public function getCurrentTimestampSQL(): string
     {
         return 'toUnixTimestamp(now())';
     }
@@ -1032,7 +856,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getListDatabasesSQL() : string
+    public function getListDatabasesSQL(): string
     {
         return 'SHOW DATABASES';
     }
@@ -1040,26 +864,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getListTableColumnsSQL($table, $database = null) : string
-    {
-        return sprintf(
-            'DESCRIBE TABLE %s',
-            ($database ? $this->quoteSingleIdentifier($database) . '.' : '') . $this->quoteSingleIdentifier($table)
-        );
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getListTablesSQL() : string
-    {
-        return "SELECT database, name FROM system.tables WHERE database != 'system' AND engine != 'View'";
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getListViewsSQL($database) : string
+    public function getListViewsSQL(string $database): string
     {
         return "SELECT name FROM system.tables WHERE database != 'system' AND engine = 'View'";
     }
@@ -1067,7 +872,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCreateViewSQL($name, $sql) : string
+    public function getCreateViewSQL(string $name, string $sql): string
     {
         return 'CREATE VIEW ' . $this->quoteIdentifier($name) . ' AS ' . $sql;
     }
@@ -1075,7 +880,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDropViewSQL($name) : string
+    public function getDropViewSQL(string $name): string
     {
         return 'DROP TABLE ' . $this->quoteIdentifier($name);
     }
@@ -1083,68 +888,66 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCreateDatabaseSQL($database) : string
+    public function getCreateDatabaseSQL(string $name): string
     {
-        return 'CREATE DATABASE ' . $database;
+        return 'CREATE DATABASE ' . $name;
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateTimeTypeDeclarationSQL(array $fieldDeclaration) : string
+    public function getDateTimeTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE_TIME, $fieldDeclaration);
+        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE_TIME, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateTimeTzTypeDeclarationSQL(array $fieldDeclaration) : string
+    public function getDateTimeTzTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE_TIME, $fieldDeclaration);
+        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE_TIME, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getTimeTypeDeclarationSQL(array $fieldDeclaration) : string
+    public function getTimeTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(StringClickHouseType::TYPE_STRING, $fieldDeclaration);
+        return $this->prepareDeclarationSQL(StringableClickHouseType::TYPE_STRING, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateTypeDeclarationSQL(array $fieldDeclaration) : string
+    public function getDateTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE, $fieldDeclaration);
+        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getFloatDeclarationSQL(array $fieldDeclaration) : string
+    public function getFloatDeclarationSQL(array $column): string
     {
         return $this->prepareDeclarationSQL(
             NumericalClickHouseType::TYPE_FLOAT . BitNumericalClickHouseType::SIXTY_FOUR_BIT,
-            $fieldDeclaration
+            $column
         );
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDefaultTransactionIsolationLevel() : int
+    public function getDefaultTransactionIsolationLevel(): TransactionIsolationLevel
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
-
-    /* supports*() methods */
 
     /**
      * {@inheritDoc}
      */
-    public function supportsTransactions() : bool
+    public function supportsSavepoints(): bool
     {
         return false;
     }
@@ -1152,46 +955,15 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function supportsSavepoints() : bool
-    {
-        return false;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function supportsPrimaryConstraints() : bool
-    {
-        return false;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function supportsForeignKeyConstraints() : bool
-    {
-        return false;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function supportsGettingAffectedRows() : bool
-    {
-        return false;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    protected function doModifyLimitQuery($query, $limit, $offset) : string
+    protected function doModifyLimitQuery(string $query, ?int $limit, int $offset): string
     {
         if ($limit === null) {
             return $query;
         }
 
         $query .= ' LIMIT ';
-        if ($offset !== null) {
+
+        if ($offset > 0) {
             $query .= $offset . ', ';
         }
 
@@ -1203,15 +975,15 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getEmptyIdentityInsertSQL($tableName, $identifierColumnName) : string
+    public function getEmptyIdentityInsertSQL(string $quotedTableName, string $quotedIdentifierColumnName): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getTruncateTableSQL($tableName, $cascade = false) : string
+    public function getTruncateTableSQL(string $tableName, bool $cascade = false): string
     {
         /**
          * For MergeTree* engines may be done with next workaround:
@@ -1219,51 +991,45 @@ class ClickHousePlatform extends AbstractPlatform
          * SELECT partition FROM system.parts WHERE table= '$tableName';
          * ALTER TABLE $tableName DROP PARTITION $partitionName
          */
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function createSavePoint($savepoint) : string
+    public function createSavePoint(string $savepoint): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function releaseSavePoint($savepoint) : string
+    public function releaseSavePoint(string $savepoint): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function rollbackSavePoint($savepoint) : string
+    public function rollbackSavePoint(string $savepoint): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw NotSupported::new(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function getReservedKeywordsClass() : string
+    public function getDefaultValueDeclarationSQL(array $column): string
     {
-        return ClickHouseKeywords::class;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getDefaultValueDeclarationSQL($field) : string
-    {
-        if (! isset($field['default'])) {
+        if (!isset($column['default'])) {
             return '';
         }
-        $defaultValue = $this->quoteStringLiteral($field['default']);
-        $fieldType    = $field['type'] ?: null;
+
+        $defaultValue = $this->quoteStringLiteral($column['default']);
+        $fieldType    = $column['type'] ?: null;
+
         if ($fieldType !== null) {
             if ($fieldType === DatableClickHouseType::TYPE_DATE ||
                 $fieldType instanceof DateType ||
@@ -1277,9 +1043,9 @@ class ClickHousePlatform extends AbstractPlatform
                     && Type::getType('BigInt')->getBindingType() === ParameterType::INTEGER
                 )
             ) {
-                $defaultValue = $field['default'];
+                $defaultValue = $column['default'];
             } elseif ($fieldType === DatableClickHouseType::TYPE_DATE_TIME &&
-                $field['default'] === $this->getCurrentTimestampSQL()
+                $column['default'] === $this->getCurrentTimestampSQL()
             ) {
                 $defaultValue = $this->getCurrentTimestampSQL();
             }
@@ -1291,42 +1057,67 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDoctrineTypeMapping($dbType) : string
+    public function getDoctrineTypeMapping(string $dbType): string
     {
-        // FixedString
-        if (stripos($dbType, 'fixedstring') === 0) {
+        if (mb_stripos($dbType, 'fixedstring') === 0) {
             $dbType = 'fixedstring';
         }
 
-        //Enum8
-        if (stripos($dbType, 'enum8') === 0) {
+        if (mb_stripos($dbType, 'enum8') === 0) {
             $dbType = 'enum8';
         }
 
-        //Enum16
-        if (stripos($dbType, 'enum16') === 0) {
+        if (mb_stripos($dbType, 'enum16') === 0) {
             $dbType = 'enum16';
         }
+
         return parent::getDoctrineTypeMapping($dbType);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function quoteStringLiteral($str) : string
+    public function quoteStringLiteral(string $str): string
     {
-        $c = $this->getStringLiteralQuoteCharacter();
-
-        return $c . addslashes($str) . $c;
+        return parent::quoteStringLiteral(addslashes($str));
     }
 
     /**
      * {@inheritDoc}
      */
-    public function quoteSingleIdentifier($str) : string
+    public function quoteSingleIdentifier(string $str): string
     {
-        $c = $this->getIdentifierQuoteCharacter();
+        return parent::quoteSingleIdentifier(addslashes($str));
+    }
 
-        return $c . addslashes($str) . $c;
+    /**
+     * {@inheritDoc}
+     */
+    public function getCurrentDatabaseExpression(): string
+    {
+        return 'DATABASE()';
+    }
+
+    protected function getDateArithmeticIntervalExpression(string $date, string $operator, string $interval, DateIntervalUnit $unit): string
+    {
+        $operation = '+' === $operator ? 'date_add' : 'date_sub';
+        $toDateFunction = strlen($date) > 10 ? 'toDateTime' : 'toDate';
+
+        return \sprintf('%s(%s, %d, %s(\'%s\'))', $operation, $unit->value, (int) $interval, $toDateFunction, $date);
+    }
+
+    public function getSetTransactionIsolationSQL(TransactionIsolationLevel $level): string
+    {
+        throw NotSupported::new(__METHOD__);
+    }
+
+    protected function createReservedKeywordsList(): KeywordList
+    {
+        return new ClickHouseKeywords();
+    }
+
+    public function createSchemaManager(Connection $connection): AbstractSchemaManager
+    {
+        return new ClickHouseSchemaManager($connection, $this);
     }
 }
