@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /*
  * This file is part of the FODDBALClickHouse package -- Doctrine DBAL library
  * for ClickHouse (a column-oriented DBMS for OLAP <https://clickhouse.yandex/>)
@@ -11,8 +14,8 @@
 
 namespace FOD\DBALClickHouse\Tests;
 
-use Doctrine\DBAL\Driver\ServerInfoAwareConnection;
-use FOD\DBALClickHouse\ClickHouseException;
+use Doctrine\DBAL\Platforms\Exception\NotSupported;
+use Doctrine\DBAL\TransactionIsolationLevel;
 use FOD\DBALClickHouse\Connection;
 use PHPUnit\Framework\TestCase;
 
@@ -23,141 +26,150 @@ use PHPUnit\Framework\TestCase;
  */
 class ConnectionTest extends TestCase
 {
-    /** @var  Connection */
-    protected $connection;
+    private Connection $connection;
 
-    public function setUp() : void
+    public function setUp(): void
     {
         $this->connection = CreateConnectionTest::createConnection();
     }
 
-    public function testExecuteUpdateDelete()
+    public function testDelete(): void
     {
-        $this->expectException(ClickHouseException::class);
-        $this->connection->executeUpdate('DELETE from test WHERE 1');
-    }
+        $this->expectException(NotSupported::class);
 
-    public function testExecuteUpdateUpdate()
-    {
-        $this->expectException(ClickHouseException::class);
-        $this->connection->executeUpdate('UPDATE test SET name = :name WHERE id = :id', [':name' => 'test', ':id' => 1]);
-    }
-
-    public function testDelete()
-    {
-        $this->expectException(ClickHouseException::class);
         $this->connection->delete('test', ['id' => 1]);
     }
 
-    public function testUpdate()
+    public function testUpdate(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->update('test', ['name' => 'test'], ['id' => 1]);
     }
 
-    public function testSetTransactionIsolation()
+    public function testExecuteStatementDelete(): void
     {
-        $this->expectException(ClickHouseException::class);
-        $this->connection->setTransactionIsolation(1);
+        $this->expectException(NotSupported::class);
+
+        $this->connection->executeStatement('DELETE FROM test WHERE id = :id', ['id' => 1]);
     }
 
-    public function testGetTransactionIsolation()
+    public function testExecuteStatementUpdate(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
+        $this->connection->executeStatement('UPDATE test SET name = :name WHERE id = :id', ['name' => 'test', 'id' => 1]);
+    }
+
+    public function testSetTransactionIsolation(): void
+    {
+        $this->expectException(NotSupported::class);
+
+        $this->connection->setTransactionIsolation(TransactionIsolationLevel::READ_COMMITTED);
+    }
+
+    public function testGetTransactionIsolation(): void
+    {
+        $this->expectException(NotSupported::class);
+
         $this->connection->getTransactionIsolation();
     }
 
-    public function testGetTransactionNestingLevel()
+    public function testGetTransactionNestingLevel(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->getTransactionNestingLevel();
     }
 
-    public function testTransactional()
+    public function testTransactional(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->transactional(function () {
         });
     }
 
-    public function testSetNestTransactionsWithSavepoints()
+    public function testSetNestTransactionsWithSavepoints(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->setNestTransactionsWithSavepoints(true);
     }
 
-    public function testGetNestTransactionsWithSavepoints()
+    public function testGetNestTransactionsWithSavepoints(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->getNestTransactionsWithSavepoints();
     }
 
-    public function testBeginTransaction()
+    public function testBeginTransaction(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->beginTransaction();
     }
 
-    public function testCommit()
+    public function testCommit(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->commit();
     }
 
-    public function testRollBack()
+    public function testRollBack(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->rollBack();
     }
 
-    public function testCreateSavepoint()
+    public function testCreateSavepoint(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->createSavepoint('1');
     }
 
-    public function testReleaseSavepoint()
+    public function testReleaseSavepoint(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->releaseSavepoint('1');
     }
 
-    public function testRollbackSavepoint()
+    public function testRollbackSavepoint(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->rollbackSavepoint('1');
     }
 
-    public function testSetRollbackOnly()
+    public function testSetRollbackOnly(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->setRollbackOnly();
     }
 
-    public function testIsRollbackOnly()
+    public function testIsRollbackOnly(): void
     {
-        $this->expectException(ClickHouseException::class);
+        $this->expectException(NotSupported::class);
+
         $this->connection->isRollbackOnly();
     }
 
-    public function testPing()
+    public function testGetServerVersion(): void
     {
-        $this->assertTrue($this->connection->ping());
-    }
+        $conn = $this->connection->getNativeConnection();
 
-    public function testGetServerVersion()
-    {
-        $conn = $this->connection->getWrappedConnection();
-        if ($conn instanceof ServerInfoAwareConnection) {
-            $pattern = '/(^[0-9]+.[0-9]+.[0-9]+(.[0-9]$|$))/mi';
-            if (method_exists($this, 'assertMatchesRegularExpression')) {
-                $this->assertMatchesRegularExpression($pattern, $conn->getServerVersion());
-            } else {
-                $this->assertRegExp($pattern, $conn->getServerVersion());
-            }
+        $pattern = '/^\d+\.\d+\.\d+\.\d+$/';
+
+        if (method_exists($this, 'assertMatchesRegularExpression')) {
+            $this->assertMatchesRegularExpression($pattern, $conn->getServerVersion());
         } else {
-            $this->fail(sprintf('`%s` does not implement the `%s` interface', \get_class($conn),
-                ServerInfoAwareConnection::class));
+            $this->assertRegExp($pattern, $conn->getServerVersion());
         }
     }
 }
