@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /*
  * This file is part of the FODDBALClickHouse package -- Doctrine DBAL library
  * for ClickHouse (a column-oriented DBMS for OLAP <https://clickhouse.yandex/>)
@@ -11,9 +14,11 @@
 
 namespace FOD\DBALClickHouse\Tests;
 
-use Doctrine\DBAL\DBALException;
+use Doctrine\DBAL\Exception;
+use Doctrine\DBAL\ParameterType;
 use Doctrine\DBAL\Schema\Index;
-use Doctrine\DBAL\Types\Type;
+use Doctrine\DBAL\Schema\Table;
+use Doctrine\DBAL\Types\Types;
 use FOD\DBALClickHouse\ClickHouseSchemaManager;
 use FOD\DBALClickHouse\Connection;
 use FOD\DBALClickHouse\Types\ArrayType;
@@ -26,72 +31,85 @@ use PHPUnit\Framework\TestCase;
  */
 class CreateSchemaTest extends TestCase
 {
-    /** @var  Connection */
-    protected $connection;
+    private Connection $connection;
 
-    public function setUp() : void
+    public function setUp(): void
     {
         $this->connection = CreateConnectionTest::createConnection();
     }
 
-    public function testGetSchemaManager()
+    public function testGetSchemaManager(): void
     {
-        $this->assertInstanceOf(ClickHouseSchemaManager::class, $this->connection->getSchemaManager());
+        $this->assertInstanceOf(ClickHouseSchemaManager::class, $this->connection->createSchemaManager());
     }
 
-    public function testCreateNewTableSQL()
+    public function testCreateNewTableSQL(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_table');
 
-        $newTable->addColumn('id', 'integer', ['unsigned' => true]);
-        $newTable->addColumn('payload', 'string');
-        $newTable->addColumn('oneVal', Type::FLOAT);
-        $newTable->addColumn('twoVal', Type::DECIMAL);
-        $newTable->addColumn('flag', Type::BOOLEAN);
-        $newTable->addColumn('mask', Type::SMALLINT);
-        $newTable->addColumn('hash', 'string', ['length' => 32, 'fixed' => true]);
+        $newTable->addColumn('id', Types::INTEGER, ['unsigned' => true]);
+        $newTable->addColumn('payload', Types::STRING);
+        $newTable->addColumn('oneVal', Types::FLOAT);
+        $newTable->addColumn('twoVal', Types::DECIMAL);
+        $newTable->addColumn('flag', Types::BOOLEAN);
+        $newTable->addColumn('mask', Types::SMALLINT);
+        $newTable->addColumn('hash', Types::STRING, ['length' => 32, 'fixed' => true]);
         $newTable->setPrimaryKey(['id']);
 
-        $migrationSQLs = $fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform());
-        $this->assertEquals("CREATE TABLE test_table (EventDate Date DEFAULT today(), id UInt32, payload String, oneVal Float64, twoVal String, flag UInt8, mask Int16, hash FixedString(32)) ENGINE = ReplacingMergeTree(EventDate, (id), 8192)",
-            implode(';', $migrationSQLs));
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
+
+        $this->assertEquals(
+            "CREATE TABLE test_table (EventDate Date DEFAULT today(), id UInt32, payload String, oneVal Float64, twoVal String, flag UInt8, mask Int16, hash FixedString(32)) ENGINE = ReplacingMergeTree(EventDate, (id), 8192)",
+            implode(';', $migrationSQLs)
+        );
+
         foreach ($migrationSQLs as $sql) {
-            $this->connection->exec($sql);
+            $this->connection->executeStatement($sql);
         }
-        $this->connection->exec('DROP TABLE test_table');
+
+        $this->connection->executeStatement('DROP TABLE test_table');
     }
 
-    public function testCreateDropTable()
+    public function testCreateDropTable(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_table');
 
-        $newTable->addColumn('id', 'integer', ['unsigned' => true]);
-        $newTable->addColumn('payload', 'string');
-        $newTable->addColumn('oneVal', Type::FLOAT);
-        $newTable->addColumn('twoVal', Type::DECIMAL);
-        $newTable->addColumn('flag', Type::BOOLEAN);
-        $newTable->addColumn('mask', Type::SMALLINT);
-        $newTable->addColumn('hash', 'string', ['length' => 32, 'fixed' => true]);
+        $newTable->addColumn('id', Types::INTEGER, ['unsigned' => true]);
+        $newTable->addColumn('payload', Types::STRING);
+        $newTable->addColumn('oneVal', Types::FLOAT);
+        $newTable->addColumn('twoVal', Types::DECIMAL);
+        $newTable->addColumn('flag', Types::BOOLEAN);
+        $newTable->addColumn('mask', Types::SMALLINT);
+        $newTable->addColumn('hash', Types::STRING, ['length' => 32, 'fixed' => true]);
         $newTable->setPrimaryKey(['id']);
 
-        foreach ($fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform()) as $sql) {
-            $this->connection->exec($sql);
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
+
+        foreach ($migrationSQLs as $sql) {
+            $this->connection->executeStatement($sql);
         }
 
-        $this->connection->exec('DROP TABLE test_table');
-        $this->expectException(DBALException::class);
-        $this->connection->exec('DROP TABLE test_table');
+        $this->connection->executeStatement('DROP TABLE test_table');
+        $this->expectException(Exception::class);
+        $this->connection->executeStatement('DROP TABLE test_table');
     }
 
-    public function testIndexGranularityOption()
+    public function testIndexGranularityOption(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_table');
@@ -101,19 +119,27 @@ class CreateSchemaTest extends TestCase
         $newTable->setPrimaryKey(['id']);
         $newTable->addOption('indexGranularity', 4096);
 
-        $migrationSQLs = $fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform());
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
         $generatedSQL = implode(';', $migrationSQLs);
-        $this->assertEquals("CREATE TABLE test_table (EventDate Date DEFAULT today(), id UInt32, payload String) ENGINE = ReplacingMergeTree(EventDate, (id), 4096)",
-            $generatedSQL);
+
+        $this->assertEquals(
+            "CREATE TABLE test_table (EventDate Date DEFAULT today(), id UInt32, payload String) ENGINE = ReplacingMergeTree(EventDate, (id), 4096)",
+            $generatedSQL
+        );
+
         foreach ($migrationSQLs as $sql) {
-            $this->connection->exec($sql);
+            $this->connection->executeStatement($sql);
         }
-        $this->connection->exec('DROP TABLE test_table');
+
+        $this->connection->executeStatement('DROP TABLE test_table');
     }
 
-    public function testEngineMergeOption()
+    public function testEngineMergeOption(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_table');
@@ -123,19 +149,27 @@ class CreateSchemaTest extends TestCase
         $newTable->setPrimaryKey(['id']);
         $newTable->addOption('engine', 'MergeTree');
 
-        $migrationSQLs = $fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform());
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
         $generatedSQL = implode(';', $migrationSQLs);
-        $this->assertEquals("CREATE TABLE test_table (EventDate Date DEFAULT today(), id UInt32, payload String) ENGINE = MergeTree(EventDate, (id), 8192)",
-            $generatedSQL);
+
+        $this->assertEquals(
+            "CREATE TABLE test_table (EventDate Date DEFAULT today(), id UInt32, payload String) ENGINE = MergeTree(EventDate, (id), 8192)",
+            $generatedSQL
+        );
+
         foreach ($migrationSQLs as $sql) {
-            $this->connection->exec($sql);
+            $this->connection->executeStatement($sql);
         }
-        $this->connection->exec('DROP TABLE test_table');
+
+        $this->connection->executeStatement('DROP TABLE test_table');
     }
 
-    public function testEngineMemoryOption()
+    public function testEngineMemoryOption(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_table');
@@ -145,18 +179,24 @@ class CreateSchemaTest extends TestCase
         $newTable->setPrimaryKey(['id']);
         $newTable->addOption('engine', 'Memory');
 
-        $migrationSQLs = $fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform());
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
         $generatedSQL = implode(';', $migrationSQLs);
+
         $this->assertEquals("CREATE TABLE test_table (id UInt32, payload String) ENGINE = Memory", $generatedSQL);
+
         foreach ($migrationSQLs as $sql) {
-            $this->connection->exec($sql);
+            $this->connection->executeStatement($sql);
         }
-        $this->connection->exec('DROP TABLE test_table');
+
+        $this->connection->executeStatement('DROP TABLE test_table');
     }
 
-    public function testEventDateColumnOption()
+    public function testEventDateColumnOption(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_table');
@@ -167,105 +207,143 @@ class CreateSchemaTest extends TestCase
         $newTable->addOption('eventDateColumn', 'event_date');
         $newTable->setPrimaryKey(['id']);
 
-        $migrationSQLs = $fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform());
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
         $generatedSQL = implode(';', $migrationSQLs);
-        $this->assertEquals("CREATE TABLE test_table (event_date Date DEFAULT today(), id UInt32, payload String) ENGINE = ReplacingMergeTree(event_date, (id), 8192)",
-            $generatedSQL);
+
+        $this->assertEquals(
+            "CREATE TABLE test_table (event_date Date DEFAULT today(), id UInt32, payload String) ENGINE = ReplacingMergeTree(event_date, (id), 8192)",
+            $generatedSQL
+        );
+
         foreach ($migrationSQLs as $sql) {
-            $this->connection->exec($sql);
+            $this->connection->executeStatement($sql);
         }
-        $this->connection->exec('DROP TABLE test_table');
+
+        $this->connection->executeStatement('DROP TABLE test_table');
     }
 
-    public function testEventDateColumnBadOption()
+    public function testEventDateColumnBadOption(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_table');
 
         $newTable->addColumn('id', 'integer', ['unsigned' => true]);
         $newTable->addColumn('payload', 'string');
-        $newTable->addColumn('event_date', Type::DATETIME, ['default' => 'toDate(now())']);
+        $newTable->addColumn('event_date', Types::DATETIME_MUTABLE, ['default' => 'toDate(now())']);
         $newTable->addOption('eventDateColumn', 'event_date');
         $newTable->setPrimaryKey(['id']);
 
         $this->expectException(\Exception::class);
-        $migrationSQLs = $fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform());
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
         $generatedSQL = implode(';', $migrationSQLs);
-        $this->assertEquals("CREATE TABLE test_table (event_date Date DEFAULT today(), id UInt32, payload String) ENGINE = ReplacingMergeTree(event_date, (id), 8192)",
-            $generatedSQL);
+
+        $this->assertEquals(
+            "CREATE TABLE test_table (event_date Date DEFAULT today(), id UInt32, payload String) ENGINE = ReplacingMergeTree(event_date, (id), 8192)",
+            $generatedSQL
+        );
+
         foreach ($migrationSQLs as $sql) {
-            $this->connection->exec($sql);
+            $this->connection->executeStatement($sql);
         }
-        $this->connection->exec('DROP TABLE test_table');
+
+        $this->connection->executeStatement('DROP TABLE test_table');
     }
 
-    public function testEventDateProviderColumnOption()
+    public function testEventDateProviderColumnOption(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_table');
 
         $newTable->addColumn('id', 'integer', ['unsigned' => true]);
         $newTable->addColumn('payload', 'string');
-        $newTable->addColumn('updated_at', Type::DATETIME);
+        $newTable->addColumn('updated_at', Types::DATETIME_MUTABLE);
         $newTable->addOption('eventDateProviderColumn', 'updated_at');
         $newTable->setPrimaryKey(['id']);
 
-        $migrationSQLs = $fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform());
+
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
+
         $generatedSQL = implode(';', $migrationSQLs);
-        $this->assertEquals("CREATE TABLE test_table (EventDate Date DEFAULT toDate(updated_at), id UInt32, payload String, updated_at DateTime) ENGINE = ReplacingMergeTree(EventDate, (id), 8192)",
-            $generatedSQL);
+
+        $this->assertEquals(
+            "CREATE TABLE test_table (EventDate Date DEFAULT toDate(updated_at), id UInt32, payload String, updated_at DateTime) ENGINE = ReplacingMergeTree(EventDate, (id), 8192)",
+            $generatedSQL
+        );
+
         foreach ($migrationSQLs as $sql) {
-            $this->connection->exec($sql);
+            $this->connection->executeStatement($sql);
         }
-        $this->connection->exec('DROP TABLE test_table');
+
+        $this->connection->executeStatement('DROP TABLE test_table');
     }
 
-    public function testEventDateProviderColumnBadOption()
+    public function testEventDateProviderColumnBadOption(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_table');
 
         $newTable->addColumn('id', 'integer', ['unsigned' => true]);
         $newTable->addColumn('payload', 'string');
-        $newTable->addColumn('flag', Type::BOOLEAN);
+        $newTable->addColumn('flag', Types::BOOLEAN);
         $newTable->addOption('eventDateProviderColumn', 'flag');
         $newTable->setPrimaryKey(['id']);
 
         $this->expectException(\Exception::class);
-        $migrationSQLs = $fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform());
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
         $generatedSQL = implode(';', $migrationSQLs);
-        $this->assertEquals("CREATE TABLE test_table (EventDate Date DEFAULT toDate(updated_at), id UInt32, payload String, updated_at DateTime) ENGINE = ReplacingMergeTree(EventDate, (id), 8192)",
-            $generatedSQL);
+
+        $this->assertEquals(
+            "CREATE TABLE test_table (EventDate Date DEFAULT toDate(updated_at), id UInt32, payload String, updated_at DateTime) ENGINE = ReplacingMergeTree(EventDate, (id), 8192)",
+            $generatedSQL
+        );
+
         foreach ($migrationSQLs as $sql) {
-            $this->connection->exec($sql);
+            $this->connection->executeStatement($sql);
         }
-        $this->connection->exec('DROP TABLE test_table');
+
+        $this->connection->executeStatement('DROP TABLE test_table');
     }
 
-    public function testListTableIndexes()
+    public function testListTableIndexes(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_indexes_table');
 
         $newTable->addColumn('id', 'integer', ['unsigned' => true]);
         $newTable->addColumn('payload', 'string');
-        $newTable->addColumn('event_date', Type::DATE);
+        $newTable->addColumn('event_date', Types::DATE_MUTABLE);
         $newTable->addOption('eventDateColumn', 'event_date');
         $newTable->setPrimaryKey(['id', 'event_date']);
-        $migrationSQLs = $fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform());
+
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
+
         foreach ($migrationSQLs as $sql) {
-            $this->connection->exec($sql);
+            $this->connection->executeStatement($sql);
         }
 
-        $indexes = $this->connection->getSchemaManager()->listTableIndexes('test_indexes_table');
+        $indexes = $this->connection->createSchemaManager()->listTableIndexes('test_indexes_table');
 
         $this->assertEquals(1, \count($indexes));
 
@@ -276,31 +354,40 @@ class CreateSchemaTest extends TestCase
             $this->assertTrue($index->isPrimary());
         }
 
-        $this->connection->exec('DROP TABLE test_indexes_table');
+        $this->connection->executeStatement('DROP TABLE test_indexes_table');
     }
 
-    public function testTableWithSamplingExpression()
+    public function testTableWithSamplingExpression(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_sampling_table');
 
         $newTable->addColumn('id', 'integer', ['unsigned' => true]);
         $newTable->addColumn('payload', 'string');
-        $newTable->addColumn('event_date', Type::DATE);
+        $newTable->addColumn('event_date', Types::DATE_MUTABLE);
         $newTable->addOption('eventDateColumn', 'event_date');
         $newTable->addOption('samplingExpression', 'intHash32(id)');
         $newTable->setPrimaryKey(['id', 'event_date']);
-        $migrationSQLs = $fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform());
+
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
+
         $generatedSQL = implode(';', $migrationSQLs);
-        $this->assertEquals("CREATE TABLE test_sampling_table (event_date Date DEFAULT today(), id UInt32, payload String) ENGINE = ReplacingMergeTree(event_date, intHash32(id), (id, event_date, intHash32(id)), 8192)",
-            $generatedSQL);
+
+        $this->assertEquals(
+            "CREATE TABLE test_sampling_table (event_date Date DEFAULT today(), id UInt32, payload String) ENGINE = ReplacingMergeTree(event_date, intHash32(id), (id, event_date, intHash32(id)), 8192)",
+            $generatedSQL
+        );
+
         foreach ($migrationSQLs as $sql) {
-            $this->connection->exec($sql);
+            $this->connection->executeStatement($sql);
         }
 
-        $indexes = $this->connection->getSchemaManager()->listTableIndexes('test_sampling_table');
+        $indexes = $this->connection->createSchemaManager()->listTableIndexes('test_sampling_table');
 
         $this->assertEquals(1, \count($indexes));
 
@@ -311,17 +398,18 @@ class CreateSchemaTest extends TestCase
             $this->assertTrue($index->isPrimary());
         }
 
-        $this->connection->exec('DROP TABLE test_sampling_table');
+        $this->connection->executeStatement('DROP TABLE test_sampling_table');
     }
 
-    public function testNullableColumns()
+    public function testNullableColumns(): void
     {
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
-        ArrayType::registerArrayTypes($this->connection->getDatabasePlatform());
-
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_table_nullable');
+
+        ArrayType::registerArrayTypes($this->connection->getDatabasePlatform());
 
         $newTable->addColumn('id', 'integer', ['unsigned' => true, 'notnull' => false]);
         $newTable->addColumn('payload', 'string', ['notnull' => false]);
@@ -331,57 +419,82 @@ class CreateSchemaTest extends TestCase
         $newTable->setPrimaryKey(['id']);
         $newTable->addOption('engine', 'Memory');
 
-        $migrationSQLs = $fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform());
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
+
         $generatedSQL = implode(';', $migrationSQLs);
-        $this->assertEquals("CREATE TABLE test_table_nullable (id UInt32, payload Nullable(String), price Nullable(Float64), transactions Array(Nullable(DateTime)), status Nullable(UInt8)) ENGINE = Memory",
-            $generatedSQL);
+
+        $this->assertEquals(
+            "CREATE TABLE test_table_nullable (id UInt32, payload Nullable(String), price Nullable(Float64), transactions Array(Nullable(DateTime)), status Nullable(UInt8)) ENGINE = Memory",
+            $generatedSQL
+        );
+
         foreach ($migrationSQLs as $sql) {
-            $this->connection->exec($sql);
+            $this->connection->executeStatement($sql);
         }
-        $this->connection->insert('test_table_nullable',
+
+        $this->connection->insert(
+            'test_table_nullable',
             [
-                'id' => 1,
-                'payload' => 's1',
-                'price' => 1.5,
+                'id'           => 1,
+                'payload'      => 's1',
+                'price'        => 1.5,
                 'transactions' => [date('Y-m-d H:i:s'), null],
-                'status' => null
-            ]);
-        $this->connection->insert('test_table_nullable',
+                'status'       => null,
+            ],
+            ['status' => ParameterType::NULL]
+        );
+        $this->connection->insert(
+            'test_table_nullable',
             [
-                'id' => 2,
-                'payload' => 's2',
-                'price' => 120,
+                'id'           => 2,
+                'payload'      => 's2',
+                'price'        => 120,
                 'transactions' => [null, null],
-                'status' => false
-            ]);
-        $this->connection->insert('test_table_nullable',
+                'status'       => false,
+            ],
+            ['status' => ParameterType::BOOLEAN]
+        );
+        $this->connection->insert(
+            'test_table_nullable',
             [
-                'id' => 3,
-                'payload' => null,
-                'price' => 1000,
+                'id'           => 3,
+                'payload'      => null,
+                'price'        => 1000,
                 'transactions' => [date('Y-m-d H:i:s')],
-                'status' => true
-            ]);
-        $this->connection->insert('test_table_nullable',
+                'status'       => true,
+            ],
+            ['status' => ParameterType::BOOLEAN]
+        );
+        $this->connection->insert(
+            'test_table_nullable',
             [
-                'id' => 4,
-                'payload' => 's4',
-                'price' => null,
+                'id'           => 4,
+                'payload'      => 's4',
+                'price'        => null,
                 'transactions' => [date('Y-m-d H:i:s'), date('Y-m-d H:i:s')],
-                'status' => null
-            ]);
-        $this->connection->insert('test_table_nullable',
+                'status'       => null,
+            ],
+            ['status' => ParameterType::NULL]
+        );
+        $this->connection->insert(
+            'test_table_nullable',
             [
-                'id' => 5,
-                'payload' => 's5',
-                'price' => 100,
+                'id'           => 5,
+                'payload'      => 's5',
+                'price'        => 100,
                 'transactions' => [date('Y-m-d H:i:s')],
-                'status' => true
-            ]);
+                'status'       => true,
+            ],
+            ['status' => ParameterType::BOOLEAN]
+        );
 
-        $this->assertEquals(2,
-            (int)$this->connection->fetchColumn("SELECT count() from test_table_nullable WHERE {$this->connection->getDatabasePlatform()->getIsNullExpression('status')}"));
+        $this->assertEquals(
+            2,
+            (int) $this->connection->fetchOne("SELECT count() from test_table_nullable WHERE isNULL(status)")
+        );
 
-        $this->connection->exec('DROP TABLE test_table_nullable');
+        $this->connection->executeStatement('DROP TABLE test_table_nullable');
     }
 }
