@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /*
  * This file is part of the FODDBALClickHouse package -- Doctrine DBAL library
  * for ClickHouse (a column-oriented DBMS for OLAP <https://clickhouse.yandex/>)
@@ -11,7 +14,7 @@
 
 namespace FOD\DBALClickHouse\Tests;
 
-use Doctrine\DBAL\FetchMode;
+use Doctrine\DBAL\ParameterType;
 use Doctrine\DBAL\Platforms\TrimMode;
 use FOD\DBALClickHouse\Connection;
 use PHPUnit\Framework\TestCase;
@@ -23,14 +26,14 @@ use PHPUnit\Framework\TestCase;
  */
 class SelectTest extends TestCase
 {
-    /** @var  Connection */
-    protected $connection;
+    private Connection $connection;
 
-    public function setUp() : void
+    public function setUp(): void
     {
         $this->connection = CreateConnectionTest::createConnection();
 
-        $fromSchema = $this->connection->getSchemaManager()->createSchema();
+        $comparator = $this->connection->createSchemaManager()->createComparator();
+        $fromSchema = $this->connection->createSchemaManager()->introspectSchema();
         $toSchema = clone $fromSchema;
 
         $newTable = $toSchema->createTable('test_select_table');
@@ -41,63 +44,58 @@ class SelectTest extends TestCase
         $newTable->addOption('engine', 'Memory');
         $newTable->setPrimaryKey(['id']);
 
-        foreach ($fromSchema->getMigrateToSql($toSchema, $this->connection->getDatabasePlatform()) as $sql) {
-            $this->connection->exec($sql);
+        $migrationSQLs = $this->connection->getDatabasePlatform()->getAlterSchemaSQL(
+            $comparator->compareSchemas($fromSchema, $toSchema)
+        );
+
+        foreach ($migrationSQLs as $sql) {
+            $this->connection->executeStatement($sql);
         }
 
-        $this->connection->exec("INSERT INTO test_select_table(id, payload, hits) VALUES (1, 'v1', 101), (2, 'v2', 202), (3, 'v3', 303), (4, 'v4', 404), (5, 'v4', 505), (6, '  t1   ', 606), (7, 'aat2aaa', 707)");
+        $this->connection->executeStatement("INSERT INTO test_select_table(id, payload, hits) VALUES (1, 'v1', 101), (2, 'v2', 202), (3, 'v3', 303), (4, 'v4', 404), (5, 'v4', 505), (6, '  t1   ', 606), (7, 'aat2aaa', 707)");
     }
 
-    public function tearDown() : void
+    public function tearDown(): void
     {
-        $this->connection->exec('DROP TABLE test_select_table');
+        $this->connection->executeStatement('DROP TABLE test_select_table');
     }
 
-    public function testFetchBothSelect()
+    public function testFetchBothSelect(): void
     {
         $results = [];
-        $stmt = $this->connection->query('SELECT * FROM test_select_table WHERE id = 3');
-        while ($result = $stmt->fetch()) {
-            $results[] = $result;
+
+        $result = $this->connection->executeQuery('SELECT * FROM test_select_table WHERE id = 3');
+
+        while ($row = $result->fetchAssociative()) {
+            $results[] = $row;
         }
+
         $this->assertEquals([['id' => 3, 'payload' => 'v3', 'hits' => 303]], $results);
     }
 
-    public function testFetchAssocSelect()
+    public function testFetchAssocSelect(): void
     {
         $results = [];
-        $stmt = $this->connection->query('SELECT id, hits FROM test_select_table WHERE id IN (3, 4)');
-        while ($result = $stmt->fetch(FetchMode::ASSOCIATIVE)) {
-            $results[] = $result;
+
+        $result = $this->connection->executeQuery('SELECT id, hits FROM test_select_table WHERE id IN (3, 4)');
+
+        while ($row = $result->fetchAssociative()) {
+            $results[] = $row;
         }
+
         $this->assertEquals([['id' => 3, 'hits' => 303], ['id' => 4, 'hits' => 404]], $results);
     }
 
-    public function testFetchNumSelect()
+    public function testFetchNumSelect():void
     {
-        $stmt = $this->connection->query('SELECT MAX(hits) as maxHits FROM test_select_table');
-        $result = $stmt->fetch(FetchMode::ASSOCIATIVE);
-        $this->assertEquals(['maxHits' => 707], $result);
+        $result = $this->connection->executeQuery('SELECT MAX(hits) as maxHits FROM test_select_table');
+
+        $this->assertEquals(['maxHits' => 707], $result->fetchAssociative());
     }
 
-    public function testFetchObjSelect()
+    public function testFetchAllBothSelect(): void
     {
-        $stmt = $this->connection->query('SELECT MAX(hits) as maxHits FROM test_select_table');
-        $result = $stmt->fetch(FetchMode::STANDARD_OBJECT);
-        $this->assertEquals((object)['maxHits' => 707], $result);
-    }
-
-    public function testFetchKeyPairSelect()
-    {
-        $stmt = $this->connection->query("SELECT id, hits FROM test_select_table WHERE id = 2");
-        $result = $stmt->fetch(\PDO::FETCH_KEY_PAIR);
-        $this->assertEquals([2 => 202], $result);
-    }
-
-    public function testFetchAllBothSelect()
-    {
-        $stmt = $this->connection->query("SELECT * FROM test_select_table WHERE id IN (1, 3)");
-        $result = $stmt->fetchAll();
+        $result = $this->connection->executeQuery("SELECT * FROM test_select_table WHERE id IN (1, 3)");
 
         $this->assertEquals([
             [
@@ -110,85 +108,55 @@ class SelectTest extends TestCase
                 'payload' => 'v3',
                 'hits' => 303,
             ]
-        ], $result);
+        ], $result->fetchAllAssociative());
     }
 
-    public function testFetchAllNumSelect()
+    public function testFetchAllNumSelect(): void
     {
-        $stmt = $this->connection->query("SELECT AVG(hits) FROM test_select_table");
-        $result = $stmt->fetchAll(FetchMode::NUMERIC);
+        $result = $this->connection->executeQuery("SELECT AVG(hits) FROM test_select_table");
 
-        $this->assertEquals([[404]], $result);
+        $this->assertEquals([[404]], $result->fetchAllNumeric());
     }
 
-    public function testFetchAllObjSelect()
+    public function testFetchColumnValidOffsetSelect(): void
     {
-        $stmt = $this->connection->query("SELECT * FROM test_select_table WHERE id IN (2, 4)");
-        $result = $stmt->fetchAll(FetchMode::STANDARD_OBJECT);
-
-        $this->assertEquals([
-            (object)[
-                'id' => 2,
-                'payload' => 'v2',
-                'hits' => 202,
-            ],
-            (object)[
-                'id' => 4,
-                'payload' => 'v4',
-                'hits' => 404,
-            ]
-        ], $result);
-    }
-
-    public function testFetchAllKeyPairSelect()
-    {
-        $stmt = $this->connection->query("SELECT payload, hits FROM test_select_table WHERE id IN (2, 4) ORDER BY id");
-        $result = $stmt->fetchAll(\PDO::FETCH_KEY_PAIR);
-
-        $this->assertEquals([
-            [
-                'v2' => 202,
-            ],
-            [
-                'v4' => 404,
-            ]
-        ], $result);
-    }
-
-    public function testFetchColumnValidOffsetSelect()
-    {
-        $stmt = $this->connection->query("SELECT payload, hits FROM test_select_table WHERE id > 1 ORDER BY id LIMIT 3");
         $results = [];
-        while ($result = $stmt->fetchColumn(1)) {
-            $results[] = $result;
+
+        $result = $this->connection->executeQuery("SELECT payload, hits FROM test_select_table WHERE id > 1 ORDER BY id LIMIT 3");
+
+        while ($row = $result->fetchNumeric()) {
+            $results[] = $row[1];
         }
 
         $this->assertEquals([202, 303, 404], $results);
     }
 
-    public function testFetchColumnInvalidOffsetSelect()
+    public function testFetchColumnInvalidOffsetSelect(): void
     {
-        $stmt = $this->connection->query("SELECT payload, hits FROM test_select_table WHERE id > 1 ORDER BY id");
         $results = [];
-        while ($result = $stmt->fetchColumn(2)) {
-            $results[] = $result;
+
+        $result = $this->connection->executeQuery("SELECT payload, hits FROM test_select_table WHERE id > 1 ORDER BY id");
+
+        while ($row = $result->fetchOne()) {
+            $results[] = $row;
         }
+
         $this->assertEquals(['v2', 'v3', 'v4', 'v4', '  t1   ', 'aat2aaa'], $results);
     }
 
-    public function testQueryBuilderSelect()
+    public function testQueryBuilderSelect(): void
     {
-        $qb = $this->connection->createQueryBuilder();
-        $result = $qb
+        $queryBuilder = $this->connection->createQueryBuilder();
+
+        $result = $queryBuilder
             ->select('payload, uniq(hits) as uniques')
             ->from('test_select_table')
             ->where('id > :id')
-            ->setParameter('id', 2, \PDO::PARAM_INT)
+            ->setParameter('id', 2, ParameterType::INTEGER)
             ->groupBy('payload')
             ->orderBy('payload')
             ->setMaxResults(2)
-            ->execute()
-            ->fetchAll();
+            ->fetchAllAssociative();
 
         $this->assertEquals([
             [
@@ -202,12 +170,13 @@ class SelectTest extends TestCase
         ], $result);
     }
 
-    public function testDynamicParametersSelect()
+    public function testDynamicParametersSelect(): void
     {
-        $stmt = $this->connection->prepare('SELECT payload, AVG(hits) AS avg_hits FROM test_select_table WHERE id > :id GROUP BY payload ORDER BY payload');
-
-        $stmt->bindValue('id', 3, 'integer');
-        $stmt->execute();
+        $result = $this->connection->executeQuery(
+            'SELECT payload, AVG(hits) AS avg_hits FROM test_select_table WHERE id > :id GROUP BY payload ORDER BY payload',
+            ['id' => 3],
+            ['id' => ParameterType::INTEGER]
+        );
 
         $this->assertEquals([
             [
@@ -222,67 +191,61 @@ class SelectTest extends TestCase
                 'payload' => 'v4',
                 'avg_hits' => 454.5,
             ],
-        ], $stmt->fetchAll());
+        ], $result->fetchAllAssociative());
     }
 
-    public function testColumnCount()
+    public function testColumnCount(): void
     {
-        $stmt = $this->connection->prepare('SELECT * FROM test_select_table');
-        $stmt->execute();
+        $result = $this->connection->executeQuery('SELECT * FROM test_select_table');
 
-        $this->assertEquals(3, $stmt->columnCount());
+        $this->assertEquals(3, $result->columnCount());
     }
 
-    public function testTrim()
+    public function testTrim(): void
     {
-        $stmt = $this->connection->prepare(
+        $result = $this->connection->executeQuery(
             sprintf(
                 'SELECT %s FROM test_select_table WHERE id = 6',
                 $this->connection->getDatabasePlatform()->getTrimExpression('payload')
             )
         );
-        $stmt->execute();
 
-        $this->assertEquals('t1', $stmt->fetchColumn());
+        $this->assertEquals('t1', $result->fetchOne());
     }
 
-    public function testTrimLeft()
+    public function testTrimLeft(): void
     {
-        $stmt = $this->connection->prepare(
+        $result = $this->connection->executeQuery(
             sprintf(
                 'SELECT %s FROM test_select_table WHERE id = 6',
                 $this->connection->getDatabasePlatform()->getTrimExpression('payload', TrimMode::LEADING)
             )
         );
-        $stmt->execute();
 
-        $this->assertEquals('t1   ', $stmt->fetchColumn());
+        $this->assertEquals('t1   ', $result->fetchOne());
     }
 
-    public function testTrimRight()
+    public function testTrimRight(): void
     {
-        $stmt = $this->connection->prepare(
+        $result = $this->connection->executeQuery(
             sprintf(
                 'SELECT %s FROM test_select_table WHERE id = 6',
                 $this->connection->getDatabasePlatform()->getTrimExpression('payload', TrimMode::TRAILING)
             )
         );
-        $stmt->execute();
 
-        $this->assertEquals('  t1', $stmt->fetchColumn());
+        $this->assertEquals('  t1', $result->fetchOne());
     }
 
-    public function testTrimChar()
+    public function testTrimChar(): void
     {
-        $stmt = $this->connection->prepare(
+        $result = $this->connection->executeQuery(
             sprintf(
                 'SELECT %s FROM test_select_table WHERE id = 7',
                 $this->connection->getDatabasePlatform()->getTrimExpression('payload', TrimMode::UNSPECIFIED, 'a')
             )
         );
-        $stmt->execute();
 
-        $this->assertEquals('t2', $stmt->fetchColumn());
+        $this->assertEquals('t2', $result->fetchOne());
     }
 }
-
